+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [unreleased]
 
 - Set default endpoints based on API key [#214](https://github.com/bugsnag/bugsnag-cli/pull/214)
+- Add minimum node version to the NPM package [#217](https://github.com/bugsnag/bugsnag-cli/pull/217)
+
+### Fixed
+
+- Ensure that the`--metadata` accepts key=value,key=value [#216](https://github.com/bugsnag/bugsnag-cli/pull/216)
 
 ## [3.1.1] - 2025-05-14